% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tiny-segmenter.R
\docType{package}
\name{rTinySegmenter-package}
\alias{rTinySegmenter-package}
\title{Very compact Japanese tokenizer}
\description{
R version of TinySegmenter, compact Japanese tokenizer
}
\details{
rTinySegmenter is a R version of TinySegmenter, which is an extremely compact Japanese tokenizer originally written in JavaScript by Mr. Taku Kudo.
}
\author{
Author: Yoshiaki AMANO. \cr
Maintainer: Yoshiaki AMANO. \cr
Twitter: @yamano357
}
\seealso{
\code{\link[openNLP:Parse_Annotator]{openNLP::Parse_Annotator}} \cr
\code{\link[NLP:annotators]{NLP::annotators}} \cr
\code{\link[tm:tokenizer]{tm::tokenizer}} \cr
\url{http://chasen.org/~taku/software/TinySegmenter/}
}
\keyword{documentation}

