% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tiny-segmenter.R
\docType{methods}
\name{tokenize}
\alias{tokenize}
\title{Tokenizing Japanese text}
\usage{
\S4method{tokenize}{tiny_segmenter}(text)
}
\arguments{
\item{\code{text}}{input single character vector}
}
\value{
\describe{segment Japanese text}
}
\description{
Tokenizing Japanese text
}
\details{
Given a `text` string, `tokenize` attempts to segment using the TinySegmenter algorithm.
}
\examples{
library(rTinySegmenter)
segmentr <- tiny_segmenter$new(text = "私の名前は中野です")
tokenized_text <- segmentr$tokenize()
# feature
segmentr$model_attr$unit
# feature value
segmentr$model_attr$score
}

